+++ NEWS.md
@@ -1,6 +1,6 @@
 # DNAcopy 1.75.3 (06/29/2023)
 
-* Added a init.c to register native (Fortran) routines and to disable symbol search
+* Added init.c to register native (Fortran) routines and to disable symbol search
 * Gzipped cytoBand.tab and converted default.DNAcopy.bdry to rda file in data directory
 
 # DNAcopy 1.75.2 (06/27/2023)